Resolve /WX errors on VC-WIN64
Currently the /WX (treat warnings as errors) compiler option can only
be used for VC-WIN32 builds but not for VC-WIN64 builds. The reason is
that there are just too many truncation warnings, like the following:

warning C4267: 'function': conversion from 'size_t' to 'int', possible loss of data
warning C4244: '=': conversion from '__int64' to 'int', possible loss of data

This is the predominant type of warning on VC-WIN64 builds and originates
from the fact that the 'int' and 'long' types are only 32bit wide, whereas
'size_t' and pointer differences require 64bit. Traditionally, the 'int'
type was used for buffer sizes, but nowadays the 'size_t' type is preferred.
Not only do these types have different sizes, but also one is signed and
the other unsigned. Since the APIs are fixed, there is nothing much one can
do about it.

This commit attempts to reduce the number of warnings. Most often it does
so by inserting the correct cast operator. Sometimes, the type of a local
variable or the signature of a private function is modified.

+++ crypto/asn1/a_int.c
@@ -200,7 +200,7 @@ static size_t c2i_ibuf(unsigned char *b, int *pneg,
 
 int i2c_ASN1_INTEGER(ASN1_INTEGER *a, unsigned char **pp)
 {
-    return i2c_ibuf(a->data, a->length, a->type & V_ASN1_NEG, pp);
+    return (int)i2c_ibuf(a->data, a->length, a->type & V_ASN1_NEG, pp);
 }
 
 /* Convert big endian buffer into uint64_t, return 0 on error */
@@ -297,7 +297,7 @@ ASN1_INTEGER *c2i_ASN1_INTEGER(ASN1_INTEGER **a, const unsigned char **pp,
     } else
         ret = *a;
 
-    if (ASN1_STRING_set(ret, NULL, r) == 0)
+    if (ASN1_STRING_set(ret, NULL, (int)r) == 0)
         goto err;
 
     c2i_ibuf(ret->data, &neg, *pp, len);
@@ -347,7 +347,7 @@ static int asn1_string_set_int64(ASN1_STRING *a, int64_t r, int itype)
         off = asn1_put_uint64(tbuf, r);
         a->type &= ~V_ASN1_NEG;
     }
-    return ASN1_STRING_set(a, tbuf + off, sizeof(tbuf) - off);
+    return ASN1_STRING_set(a, tbuf + off, (int)(sizeof(tbuf) - off));
 }
 
 static int asn1_string_get_uint64(uint64_t *pr, const ASN1_STRING *a,
@@ -375,7 +375,7 @@ static int asn1_string_set_uint64(ASN1_STRING *a, uint64_t r, int itype)
 
     a->type = itype;
     off = asn1_put_uint64(tbuf, r);
-    return ASN1_STRING_set(a, tbuf + off, sizeof(tbuf) - off);
+    return ASN1_STRING_set(a, tbuf + off, (int)(sizeof(tbuf) - off));
 }
 
 /*
@@ -625,6 +625,6 @@ int i2c_uint64_int(unsigned char *p, uint64_t r, int neg)
     size_t off;
 
     off = asn1_put_uint64(buf, r);
-    return i2c_ibuf(buf + off, sizeof(buf) - off, neg, &p);
+    return (int)i2c_ibuf(buf + off, sizeof(buf) - off, neg, &p);
 }
