+++ crypto/asn1/asn1_lib.c
@@ -277,7 +277,7 @@ int ASN1_STRING_set(ASN1_STRING *str, const void *_data, int len)
         if (data == NULL)
             return 0;
         else
-            len = strlen(data);
+            len = (int)strlen(data);
     }
     if ((str->length <= len) || (str->data == NULL)) {
         c = str->data;
