+++ crypto/dsa/dsa_locl.h
@@ -64,14 +64,14 @@ struct dsa_method {
     int (*dsa_keygen) (DSA *dsa);
 };
 
-int dsa_builtin_paramgen(DSA *ret, size_t bits, size_t qbits,
+int dsa_builtin_paramgen(DSA *ret, int bits, int qbits,
                          const EVP_MD *evpmd, const unsigned char *seed_in,
-                         size_t seed_len, unsigned char *seed_out,
+                         int seed_len, unsigned char *seed_out,
                          int *counter_ret, unsigned long *h_ret,
                          BN_GENCB *cb);
 
-int dsa_builtin_paramgen2(DSA *ret, size_t L, size_t N,
+int dsa_builtin_paramgen2(DSA *ret, int L, int N,
                           const EVP_MD *evpmd, const unsigned char *seed_in,
-                          size_t seed_len, int idx, unsigned char *seed_out,
+                          int seed_len, int idx, unsigned char *seed_out,
                           int *counter_ret, unsigned long *h_ret,
                           BN_GENCB *cb);
