+++ crypto/ec/ec_key.c
@@ -562,9 +562,9 @@ size_t EC_KEY_priv2oct(const EC_KEY *eckey,
 size_t ec_key_simple_priv2oct(const EC_KEY *eckey,
                               unsigned char *buf, size_t len)
 {
-    size_t buf_len;
+    unsigned int buf_len;
 
-    buf_len = (EC_GROUP_order_bits(eckey->group) + 7) / 8;
+    buf_len = (unsigned int)(EC_GROUP_order_bits(eckey->group) + 7) / 8;
     if (eckey->priv_key == NULL)
         return 0;
     if (buf == NULL)
@@ -601,7 +601,7 @@ int ec_key_simple_oct2priv(EC_KEY *eckey, const unsigned char *buf, size_t len)
         ECerr(EC_F_EC_KEY_SIMPLE_OCT2PRIV, ERR_R_MALLOC_FAILURE);
         return 0;
     }
-    eckey->priv_key = BN_bin2bn(buf, len, eckey->priv_key);
+    eckey->priv_key = BN_bin2bn(buf, (int)len, eckey->priv_key);
     if (eckey->priv_key == NULL) {
         ECerr(EC_F_EC_KEY_SIMPLE_OCT2PRIV, ERR_R_BN_LIB);
         return 0;
